bot/requirements.txt
PySocks==1.6.8
python-telegram-bot==10.1.0
requests==2.19.1
